backend/requirements.txt
uvicorn==0.38.0
fastapi==0.121.0
bcrypt==4.3.0
PyJWT==2.12.0
PyYAML==6.0.2
requests==2.32.4
